+++ src/gui/gt_mdiitem.h
@@ -15,7 +15,7 @@
 #include <QPointer>
 
 #include "gt_object.h"
-//#include "gt_variantproperty.h"
+#include "gt_shortcutsettingsdata.h"
 
 class QFrame;
 class QSettings;
@@ -154,6 +154,37 @@ protected:
      * @param subWin
      */
     void setSubWin(QMdiSubWindow* subWin);
+
+    /**
+     * @brief registerShortCut
+     * @param id - id of the shortcut to use
+     * @param k - key sequence to use
+     * @param readOnly - optional flag to define it as read only
+     * @return the key sequence used for this id
+     */
+    QKeySequence registerShortCut(QString const& id,
+                                  QKeySequence const& k,
+                                  bool readOnly = false);
+
+    /**
+     * @brief registerShortCut
+     * @param data - short cut data to register
+     * @return the key sequence used for this id
+     */
+    QKeySequence registerShortCut(GtShortCutSettingsData const& data);
+
+    /**
+     * @brief registerShortCuts
+     * @param list - list of short cuts to register
+     */
+    void registerShortCuts(QList<GtShortCutSettingsData> const& list);
+
+    /**
+     * @brief getShortCut
+     * @param id - identification string of the short cut to read
+     * @return short cut for this id registered in this object
+     */
+    QKeySequence getShortCut(const QString& id);
 
 private:
     ///